+++ src/Hpack/Defaults.hs
@@ -51,20 +51,21 @@ get url file = do
 formatStatus :: Status -> String
 formatStatus (Status code message) = show code ++ " " ++ B.unpack message
 
-ensure :: FilePath -> Defaults -> IO (Either String FilePath)
-ensure dir = \ case
+ensure :: FilePath -> FilePath -> Defaults -> IO (Either String FilePath)
+ensure userDataDir dir = \ case
   DefaultsGithub defaults -> do
     let
       url = defaultsUrl defaults
-      file = defaultsCachePath dir defaults
+      file = defaultsCachePath userDataDir defaults
     ensureFile file url >>= \ case
       Found -> return (Right file)
       NotFound -> return (Left $ notFound url)
       Failed err -> return (Left err)
-  DefaultsLocal (Local file) ->
-    doesFileExist file >>= \ case
-      True -> return (Right file)
-      False -> return (Left $ notFound file)
+  DefaultsLocal (Local file) -> do
+    let file' = dir </> file
+    doesFileExist file' >>= \ case
+      True -> return (Right file')
+      False -> return (Left $ notFound file')
   where
     notFound file = "Invalid value for \"defaults\"! File " ++ file ++ " does not exist!"